[Android] Wire up streaming capability with tool calling for remote inference
Note that due to the nature of how streaming events are coming in to the client side, there is no way to detect this is a tool call event at the moment until we hit delta.isToolCallDelta(). Thus, in streaming mode, we actually print out model responses while making tool calls. This is different from how we are doing for non-streaming.