Clean-up iOSCalendarAssitantWithLocalInf Project Dependencies (#206)
# What does this PR do?

This PR cleans-up the project dependency within the
iOSCalendarAssitantWithLocalInf:
 - Add LlamaStackClient (Swift SDK) project dependency
- Remove ExecuTorch dependency. As a result, remove the force-loading of
the backends which causes re-registering issues. ExecuTorch will now be
sourced from the Swift SDK only. This is be improved even further.
- Maintain reference dummy files for tokenizer, model pte, and Swift SDK
project. This can also be improved further.

Overall, this was part of the effort to be able to update the ExecuTorch
module in the Swift SDK
([PR](https://github.com/meta-llama/llama-stack-client-swift/pull/27)).

## Feature/Issue validation/testing/test plan

This has been tested with the [iOSCalendarAssistantWithLocaInfr
project](https://github.com/meta-llama/llama-stack-apps/tree/main/examples/ios_calendar_assistant)
and Llama 3.2-3B Spinquant pte + tokenizer.model. We are able to ask
"What is the capital of France" and it generates a response.